created new functions file and referenced in app.py
